提交测试

+++ jpf-domain/src/main/java/com/joiest/jpf/common/po/PayShopOrder.java
@@ -76,9 +76,9 @@ public class PayShopOrder implements Serializable {
     private BigDecimal productMoney;
 
     /**
-     * 产品豆价格
+     * 产品豆数量
      */
-    private Integer productDou;
+    private BigDecimal productDou;
 
     /**
      * 商品基础信息id
@@ -101,9 +101,9 @@ public class PayShopOrder implements Serializable {
     private BigDecimal totalMoney;
 
     /**
-     * 订单总豆数量
+     * 订单总豆
      */
-    private Integer totalDou;
+    private BigDecimal totalDou;
 
     /**
      * 支付方式 0=豆支付 1=微信支付
@@ -314,11 +314,11 @@ public class PayShopOrder implements Serializable {
         this.productMoney = productMoney;
     }
 
-    public Integer getProductDou() {
+    public BigDecimal getProductDou() {
         return productDou;
     }
 
-    public void setProductDou(Integer productDou) {
+    public void setProductDou(BigDecimal productDou) {
         this.productDou = productDou;
     }
 
@@ -354,11 +354,11 @@ public class PayShopOrder implements Serializable {
         this.totalMoney = totalMoney;
     }
 
-    public Integer getTotalDou() {
+    public BigDecimal getTotalDou() {
         return totalDou;
     }
 
-    public void setTotalDou(Integer totalDou) {
+    public void setTotalDou(BigDecimal totalDou) {
         this.totalDou = totalDou;
     }
